Mirror Chern Number example fix

+++ src/classes/@Oper/Rc2Rf.m
@@ -1,3 +1,3 @@
 function Rf = Rc2Rf(Rc,Rm)
-Rf =int8( Rm*Rc/Rm);
+Rf =( Rm*Rc/Rm);
 end
